Working Run Length Encoding for strings. Multiple implementations including the streaming encoding one chosen are in the source code.
RLE tested for ints too. But not built into the Builder/Factory yet.

The StringColumnBuilder chooses encoding based on heuristics now. It can also be
forced to pick one "scheme" or the other using ENV variables.

Dictionary - attempts were made to make it a complete decorator like RLE or Null
(but this still does not work)

An attempt to use traits-oriented decorators in scala style also did not work
out due to scala reflection troubles.

An LZF implementation is also present now for comparisons. It does not yet
support a block-based or streaming decompression yet which is neccessary to
avoid decompressing the whole column when needed.

--------

After RLE string compressions here are current comparisons

column	#rows	uncompressed 	compressed
u	250,000	8,501,818	466,466
f	250,000	6,814,436	6,814,436
d	250,000	6,001,288	56
a	250,000	10,502,242	74
id	250,000	5,501,182	54

columns d, a and id are partition columns and are the single-value degenerate
case for RLE.

column had too many transitions in values and heuristics chose to not compress
it.